bots/vector/pyproject.toml
[project]
name = "transform-vector"
version = "0.1.0"
description = "Load the file and transform to vector"
dependencies = [
    "python-dotenv~=1.0.0",
    "supabase~=1.0.4",
]
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true

[tool.hatch.metadata]
allow-direct-references = true
